@@ -1,6 +1,5 @@
-export default {
-    plugins: {
-        tailwindcss: {},
-        autoprefixer: {},
-    },
+export const plugins = {
+    '@tailwindcss': {},
+    autoprefixer: {},
+    '@tailwindcss/postcss': {},
 };
