@@ -1,40 +1,62 @@
 <?php
 
-use App\Http\Controllers\listkaryawanController;
-use App\Http\Controllers\HasilController;
-use App\Http\Controllers\NilaiKaryawanController;
+use App\Http\Controllers\WPController;
+use App\Http\Controllers\SAWController;
 use App\Http\Controllers\ProfileController;
-use App\Http\Controllers\KaryawanController;
+use Illuminate\Foundation\Application;
 use Illuminate\Support\Facades\Route;
+use Inertia\Inertia;
+use App\Http\Controllers\KaryawanController;
+
+Route::get('/rekomendasi', function () {
+    return Inertia::render('Rekomendasi');
+})->name('rekomendasi');
 
 Route::get('/', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+    return Inertia::render('Home');
+})->name('homepage');
 
-// Butuh akses admin atau manager
-Route::middleware(['role:admin,manager'])->group(function () {
-    Route::post('/addgrade', [KaryawanController::class, 'addGrade'])->name('employees.addgrade');
-    Route::get('/listkaryawan', [karyawanController::class, 'index'])->name('listkaryawan.index');
-    Route::get('/listkaryawan/create', [karyawanController::class, 'create'])->name('listkaryawan.create');
-    Route::post('/listkaryawan/store', [karyawanController::class, 'store'])->name('listkaryawan.store');
-    Route::get('/listkaryawan/{id}/edit', [karyawanController::class, 'edit'])->name('listkaryawan.edit');
-    Route::put('/listkaryawan/{id}/update', [karyawanController::class, 'update'])->name('listkaryawan.update');
-    Route::delete('/listkaryawan/{id}/destroy', [karyawanController::class, 'destroy'])->name('listkaryawan.destroy');
-    Route::get('/addgrade/{id}', [KaryawanController::class, 'showAddGradeForm'])->name('employees.showAddGradeForm');
-    Route::patch('/nilaikaryawan/{id}/{column}/update', [KaryawanController::class, 'updateGrade'])->name('nilaikaryawan.update');
-    Route::get('/listkaryawansearch', [KaryawanController::class, 'search'])->name('listkaryawansearch');
-    Route::get('/nilaikaryawansearch', [KaryawanController::class, 'search'])->name('nilaikaryawansearch');
-    Route::get('/nilaikaryawan', [KaryawanController::class, 'index'])->name('nilaikaryawan.index');
-    Route::get('/kriteria', [KaryawanController::class, 'index'])->name('kriteria');
-    Route::get('/subkriteria', [KaryawanController::class, 'index'])->name('subkriteria');
-});
+Route::get('/home', function () {
+    return Inertia::render('Home');
+})->name('homepage');
 
-Route::middleware('auth')->group(function () {
-    Route::get('/hasilakhir', [HasilController::class, 'index'])->name('hasil.index');
-    Route::get('/profile', [ProfileController::class, 'index'])->name('profile.index');
+// Route::middleware('auth', 'role:admin,manager')->group(function () {
+Route::middleware(['auth', 'role:admin,manager'])->group(function () {
+    Route::get('/dashboard', function () {
+        return Inertia::render('Home');
+    })->name('dashboard');
+
+    Route::get('/inputdata', function () {
+        return Inertia::render('DataForm');
+    })->name('inputdata');
+
+    Route::get('/matrix', function () {
+        return Inertia::render('Matrix');
+    })->name('matrix');
+
+    Route::get('/preferensi', function () {
+        return Inertia::render('Preferensi');
+    })->name('preferensi');
+
+    Route::get('/getdata', [KaryawanController::class, 'show']);
+
+    // Input Data
+    Route::post('/add', [KaryawanController::class, 'store']);
+    Route::put('/update', [KaryawanController::class, 'update']);
+    Route::delete('/{type}/{id}', [KaryawanController::class, 'destroy']);
+    Route::post('/addsubcriteria/{id}', [KaryawanController::class, 'addSubCriteria']);
+
+    Route::post('/addscore', [KaryawanController::class, 'addScore']);
+    Route::get('/getscore', [KaryawanController::class, 'getScore']);
+
+    Route::get('/saw/calculate', [SAWController::class, 'calculate'])->name('saw.calculate');
+    Route::get('/wp/calculate', [WPController::class, 'calculate'])->name('wp.calculate');
+
+    // Profile
+    Route::get('/username', [ProfileController::class, 'get']);
+    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-    Route::get('/logout', [ProfileController::class, 'logout'])->name('logout');
 });
 
 require __DIR__ . '/auth.php';
